requirements.txt
blinker
flask-dance
flask-login
flask-marshmallow
flask-sqlalchemy
marshmallow-sqlalchemy
sqlalchemy-utils
